requirements.txt
paho-mqtt==2.2.1
pillow>=9.0.0
opencv-python-headless==4.5.5.64